Add files via upload
Here is a working example of reading from 8 channels with the EEPROM disabled. I verified that this works with both ADC boards stacked (correctly read in from a potentiometer from both boards), and that changing the chip select between 3 and 8 in the code properly selects the top and bottom boards respectively (EEPROM needs to be disabled for this to work).